renamed packages and tidied up before middleware
