+++ server/models/responseModel.go
@@ -1,4 +1,4 @@
-package models
+package model
 
 type ErrorResponse struct {
 	Message string `json:"message"`
