+++ server/routes/authRouter.go
@@ -1,12 +1,12 @@
 package routes
 
-// import (
-// 	"task-manager/server/controllers"
+import (
+	controller "task-manager/server/controllers"
 
-// 	"github.com/gin-gonic/gin"
-// )
+	"github.com/gin-gonic/gin"
+)
 
-// func AuthRoutes(incomingRoutes *gin.Engine) {
-// 	incomingRoutes.POST("users/signup", controllers.Signup())
-// 	incomingRoutes.POST("users/login", controllers.Signup())
-// }
+func AuthRoutes(incomingRoutes *gin.Engine) {
+	incomingRoutes.POST("users/signup", controller.Signup())
+	incomingRoutes.POST("users/login", controller.Login())
+}
