+++ server/routes/old-routes.go
@@ -3,15 +3,15 @@ package routes
 import (
 	"github.com/gin-gonic/gin"
 
-	"task-manager/server/controllers"
+	controller "task-manager/server/controllers"
 )
 
 // SetupRoutes configures the routes and binds them to the corresponding handler functions
 func SetupRoutes(router *gin.Engine) {
-	router.GET("/tasks", controllers.GetTasks)
-	router.GET("/tasks/:id", controllers.GetTaskByID)
-	router.POST("/tasks", controllers.PostTask)
-	router.PUT("/tasks/:id", controllers.UpdateTask)
-	router.DELETE("/tasks/:id", controllers.DeleteTask)
-	router.DELETE("/tasks/", controllers.DeleteAllTasks)
+	router.GET("/tasks", controller.GetTasks)
+	router.GET("/tasks/:id", controller.GetTaskByID)
+	router.POST("/tasks", controller.PostTask)
+	router.PUT("/tasks/:id", controller.UpdateTask)
+	router.DELETE("/tasks/:id", controller.DeleteTask)
+	router.DELETE("/tasks/", controller.DeleteAllTasks)
 }
